+++ src/solvers/evolutionary_solver_cpu_impl.hpp
@@ -5,7 +5,7 @@ namespace locusta {
 template <typename TFloat>
 evolutionary_solver_cpu<TFloat>::evolutionary_solver_cpu(
   population_set_cpu<TFloat>* population, evaluator_cpu<TFloat>* evaluator,
-  prngenerator_cpu<TFloat>* prn_generator, uint32_t generation_target,
+  prngenerator_cpu<TFloat>* prn_generator, uint64_t generation_target,
   TFloat* upper_bounds, TFloat* lower_bounds)
   : evolutionary_solver<TFloat>(population, evaluator, prn_generator,
                                 generation_target, upper_bounds, lower_bounds)
